+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## 1.6.4
+* add: `parserOptions.ecmaFeatures.jsx` = true to eslint config to support jsx in `.vue` SFC(lang='tsx').`vue2-ts`,`vut3-ts`,
 ## 1.6.3
 * update: eslint.vue3-ts varsIgnorePattern: 'h'
 * update: deps version `ts` `vue` `eslint` `prettier`